+++ CHANGES.rst
@@ -4,4 +4,6 @@
 - Development script for slit tracing (dev_algorithms/dev_trace_slits.py)
 - Cooked/ folder (downloads via Google Drive)
 - Script to generate Cooked/ folder;  supersedes the Drive
+- Template files
+- New Keck LRISr 1200_900 file